Made Homework 1 Exercise B Better
Instead of printing the first half of the diamond, then the second half, program now prints the whole diamond at the same time by starting the loops at a negative index, and uses the absolute value of the index values to determine what is printed. Basically it takes advantage of the diamonds 4-way symmetry. This idea came to me and seemed much cleaner, so I just had to go back and "fix" it.