Just checking.
I had two versions of Exercise A in the works... decided to scrap the local version (never uploaded to git). Wanted to make sure it transferred here properly. layout, depend, and exe always need to update after the project is open for some reason. Also did one line of housekeeping to diamond.cpp.