Completed HW 5, improved RTI
Homework 5:
Is ready to submit.
Fixed Exercise A, it was dumb, it now uses stringstreams instead of the *100000000-(%10-/10)x8 method.
Exercise B was on outputting nested collatz into file.
Exercise C was on .cpp .hpp stuff and determining if a vector is sorted.

Real Time Input:
Still using system("CLS") and can't seem to get rid of the flickering. Might need to make a couple OS-specific functions.
Made some aesthetic changes, to work with more font types, makes it more universal.
Added some placeholder functions which may be used as in-game menus later.
Fixed the getch() stuff. It actually stores both halves of the key input as a 4-digit hex code now. This allows the distinction between arrow keys and ASCII letters, maybe more in the future.
Re-designated keys to cases. Lowercase and uppercase letters don't do the same thing anymore. Arrow keys can now be used for movement as well.
Last pressed key prints to bottom of map, for debugging and for streamlining future key assignment.